+++ test/test_helper.exs
@@ -1,3 +1,4 @@
 ExUnit.start()
 
 Mimic.copy(System)
+Mimic.copy(Severance.System.Test)